+++ tests/managed_hooks.rs
@@ -4,7 +4,7 @@ use std::sync::atomic::{AtomicUsize, Ordering};
 
 use async_trait::async_trait;
 
-use deadpool::managed::{Hook, HookError, HookErrorCause, Manager, Pool, PoolError, RecycleResult};
+use deadpool::managed::{Hook, HookError, Manager, Pool, RecycleResult};
 
 struct Computer {
     next_id: AtomicUsize,
@@ -63,40 +63,17 @@ async fn post_create_ok_async() {
 }
 
 #[tokio::test]
-async fn post_create_err_continue() {
+async fn post_create_err_abort() {
     let manager = Computer::new(0);
     let pool = Pool::<Computer>::builder(manager)
         .max_size(3)
         .post_create(Hook::sync_fn(|obj, _| {
-            if *obj % 2 == 0 {
-                Ok(())
-            } else {
-                Err(HookError::Continue(None))
-            }
+            (*obj % 2 == 0)
+                .then_some(())
+                .ok_or(HookError::StaticMessage("odd creation"))
         }))
         .build()
         .unwrap();
-    let obj1 = pool.get().await.unwrap();
-    assert_eq!(*obj1, 0);
-    let obj2 = pool.get().await.unwrap();
-    assert_eq!(*obj2, 2);
-    let obj3 = pool.get().await.unwrap();
-    assert_eq!(*obj3, 4);
-}
-
-#[tokio::test]
-async fn post_create_err_abort() {
-    let manager = Computer::new(0);
-    let pool =
-        Pool::<Computer>::builder(manager)
-            .max_size(3)
-            .post_create(Hook::sync_fn(|obj, _| {
-                (*obj % 2 == 0).then_some(()).ok_or(HookError::Abort(
-                    HookErrorCause::StaticMessage("odd creation"),
-                ))
-            }))
-            .build()
-            .unwrap();
     let obj1 = pool.get().await.unwrap();
     assert_eq!(*obj1, 0);
     assert!(pool.get().await.is_err());
@@ -131,7 +108,7 @@ async fn pre_recycle_err_continue() {
         .max_size(1)
         .pre_recycle(Hook::sync_fn(|_, metrics| {
             if metrics.recycle_count > 0 {
-                Err(HookError::Continue(None))
+                Err(HookError::StaticMessage("Fail!"))
             } else {
                 Ok(())
             }
@@ -156,62 +133,6 @@ async fn pre_recycle_err_continue() {
     assert_eq!(*pool.get().await.unwrap(), 2);
     assert_eq!(pool.status().available, 1);
     assert_eq!(pool.status().size, 1);
-}
-
-#[tokio::test]
-async fn pre_recycle_err_abort() {
-    let manager = Computer::new(0);
-    let pool = Pool::<Computer>::builder(manager)
-        .max_size(1)
-        .pre_recycle(Hook::sync_fn(|_, metrics| {
-            if metrics.recycle_count > 0 {
-                Err(HookError::Abort(HookErrorCause::StaticMessage(
-                    "no object recycling",
-                )))
-            } else {
-                Ok(())
-            }
-        }))
-        .build()
-        .unwrap();
-    assert_eq!(pool.status().available, 0);
-    assert_eq!(pool.status().size, 0);
-    assert!(matches!(pool.get().await, Ok(x) if *x == 0));
-    assert_eq!(pool.status().available, 1);
-    assert_eq!(pool.status().size, 1);
-    assert!(matches!(pool.get().await, Ok(x) if *x == 0));
-    assert_eq!(pool.status().available, 1);
-    assert_eq!(pool.status().size, 1);
-    assert!(matches!(
-        pool.get().await,
-        Err(PoolError::PreRecycleHook(HookError::Abort(_)))
-    ));
-    assert_eq!(pool.status().available, 0);
-    assert_eq!(pool.status().size, 0);
-    assert!(matches!(pool.get().await, Ok(x) if *x == 1));
-    assert_eq!(pool.status().available, 1);
-    assert_eq!(pool.status().size, 1);
-    assert!(matches!(pool.get().await, Ok(x) if *x == 1));
-    assert_eq!(pool.status().available, 1);
-    assert_eq!(pool.status().size, 1);
-    assert!(matches!(
-        pool.get().await,
-        Err(PoolError::PreRecycleHook(HookError::Abort(_)))
-    ));
-    assert_eq!(pool.status().available, 0);
-    assert_eq!(pool.status().size, 0);
-    assert!(matches!(pool.get().await, Ok(x) if *x == 2));
-    assert_eq!(pool.status().available, 1);
-    assert_eq!(pool.status().size, 1);
-    assert!(matches!(pool.get().await, Ok(x) if *x == 2));
-    assert_eq!(pool.status().available, 1);
-    assert_eq!(pool.status().size, 1);
-    assert!(matches!(
-        pool.get().await,
-        Err(PoolError::PreRecycleHook(HookError::Abort(_)))
-    ));
-    assert_eq!(pool.status().available, 0);
-    assert_eq!(pool.status().size, 0);
 }
 
 #[tokio::test]
@@ -238,7 +159,7 @@ async fn post_recycle_err_continue() {
         .max_size(1)
         .post_recycle(Hook::sync_fn(|_, metrics| {
             if metrics.recycle_count > 0 {
-                Err(HookError::Continue(None))
+                Err(HookError::StaticMessage("Fail!"))
             } else {
                 Ok(())
             }
@@ -263,60 +184,4 @@ async fn post_recycle_err_continue() {
     assert_eq!(*pool.get().await.unwrap(), 2);
     assert_eq!(pool.status().available, 1);
     assert_eq!(pool.status().size, 1);
-}
-
-#[tokio::test]
-async fn post_recycle_err_abort() {
-    let manager = Computer::new(0);
-    let pool = Pool::<Computer>::builder(manager)
-        .max_size(1)
-        .post_recycle(Hook::sync_fn(|_, metrics| {
-            if metrics.recycle_count > 0 {
-                Err(HookError::Abort(HookErrorCause::StaticMessage(
-                    "no object recycling",
-                )))
-            } else {
-                Ok(())
-            }
-        }))
-        .build()
-        .unwrap();
-    assert_eq!(pool.status().available, 0);
-    assert_eq!(pool.status().size, 0);
-    assert!(matches!(pool.get().await, Ok(x) if *x == 0));
-    assert_eq!(pool.status().available, 1);
-    assert_eq!(pool.status().size, 1);
-    assert!(matches!(pool.get().await, Ok(x) if *x == 0));
-    assert_eq!(pool.status().available, 1);
-    assert_eq!(pool.status().size, 1);
-    assert!(matches!(
-        pool.get().await,
-        Err(PoolError::PostRecycleHook(HookError::Abort(_)))
-    ));
-    assert_eq!(pool.status().available, 0);
-    assert_eq!(pool.status().size, 0);
-    assert!(matches!(pool.get().await, Ok(x) if *x == 1));
-    assert_eq!(pool.status().available, 1);
-    assert_eq!(pool.status().size, 1);
-    assert!(matches!(pool.get().await, Ok(x) if *x == 1));
-    assert_eq!(pool.status().available, 1);
-    assert_eq!(pool.status().size, 1);
-    assert!(matches!(
-        pool.get().await,
-        Err(PoolError::PostRecycleHook(HookError::Abort(_)))
-    ));
-    assert_eq!(pool.status().available, 0);
-    assert_eq!(pool.status().size, 0);
-    assert!(matches!(pool.get().await, Ok(x) if *x == 2));
-    assert_eq!(pool.status().available, 1);
-    assert_eq!(pool.status().size, 1);
-    assert!(matches!(pool.get().await, Ok(x) if *x == 2));
-    assert_eq!(pool.status().available, 1);
-    assert_eq!(pool.status().size, 1);
-    assert!(matches!(
-        pool.get().await,
-        Err(PoolError::PostRecycleHook(HookError::Abort(_)))
-    ));
-    assert_eq!(pool.status().available, 0);
-    assert_eq!(pool.status().size, 0);
 }