+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Change Log
 
-## v0.10.0
+## v0.10.0 (unreleased)
 
 - Remove unreachable enum variant `BuildError::Backend`
+- Remove `HookError::Continue` and `HookError::Abort` variants
+  replacing it with the contents of `HookErrorCause`. Returning
+  a `HookError` from a `post_create` hook causes the `Pool::get`
+  operation to fail while returning it from a `pre_recycle` or `post_recycle` hook the operation continues.
 
 ## v0.9.5